perf(rust): move temp_filter, ssr_schedule, gain_schedule to f32
The MAX31856 already returns f32, so the f64 in these three leaf modules
was widen-then-narrow with no added precision. Switch their stored state
and public API to f32:

- temp_filter: median window, offset, last-good. Bounded [-50,1500] C;
  f32 ulp there (~1e-4 C) is 64x finer than the sensor LSB (0.0078 C).
- ssr_schedule: duty % only (time is already integer ms). The ON/OFF
  on-time is still computed in f64 so the relay decision stays bit-exact.
- gain_schedule: gains are recomputed fresh each tick (no accumulation),
  and the change thresholds sit far above f32 noise.

Boundary casts are kept where pid/state/scheduler are still f64. Golden
replay tolerances relaxed (1e-9 -> 1e-3/1e-4) for the f32 paths.

Measured (stable build, real binary):
  RAM (.data+.bss): 206,980 -> 206,900 B  (control_task pool -88 B)
  Flash:            +192 B  (f32<->f64 conversions at the still-f64
                             boundary; reverses once pid/state follow)

Also fix scripts/ram.sh: an empty exported RUSTFLAGS was shadowing the
config.toml -Tlink.x linker script, so every run linked a 20-byte stub
and reported 0 B RAM. Build with the default toolchain (nightly only for
--types) and measure the final linked binary, not the newest deps/ ELF.

+++ rust/kiln-app/src/config.rs
@@ -275,7 +275,11 @@ impl KilnConfig {
 
     /// Base PID gains as the gain scheduler's triple.
     pub fn pid_base(&self) -> Gains {
-        Gains::new(self.pid_kp_base, self.pid_ki_base, self.pid_kd_base)
+        Gains::new(
+            self.pid_kp_base as f32,
+            self.pid_ki_base as f32,
+            self.pid_kd_base as f32,
+        )
     }
 
     /// Outer control-tick period in milliseconds.
